requirements.txt
numpy
tqdm
silero-vad
tokenizers
datetime
soundfile
huggingface-hub
piper-tts
websocket-server
llama-cpp-python @ https://synaptics-synap.github.io/examples-prebuilts/packages/llama_cpp_python-0.3.16-cp312-cp312-linux_aarch64.whl
synap-python @ https://github.com/synaptics-synap/synap-python/releases/download/v0.0.4-preview/synap_python-0.0.4-cp312-cp312-manylinux_2_35_aarch64.whl
onnxruntime>=1.20.0,<1.21.0
transformers==4.45.1
tokenizers<0.21
torch==2.9.0
torchaudio=2.9.0
dotenv
scikit-learn
anthropic
fugashi
sentencepiece
unidic-lite
transitions
